requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-Cors==4.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
boto3==1.34.0
SQLAlchemy==2.0.23
email-validator==2.1.0
